AC-2 create base entities

+++ database/migrations/2022_03_04_204849_create_transactions_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->id();
             $table->integer('sender_wallet_id');
             $table->integer('recipient_wallet_id');
-            $table->float('volume');
+            $table->integer('volume');
             $table->integer('status');
             $table->integer('type');
             $table->timestamps();
